adjust SearchProvider datas
