+++ constants/globalTypes.tsx
@@ -81,6 +81,7 @@ export type CitiesDataItemType = {
 };
 
 export type CitiesDataType = CitiesDataItemType[];
+
 // store
 export type StoreDataType = {
   storeId: number;
@@ -121,3 +122,25 @@ export type GroupTagIdType =
   | "competitiveTag";
 
 export type GroupTagItemType = { id: GroupTagIdType; text: string };
+
+// SearchProvider
+export type SearchProviderProps = { children: ReactNode };
+
+export type SearchValuesType = {
+  cityId: number;
+  page: number;
+  pageSize: number;
+
+  // group
+  startDate: string;
+  gameName: string;
+  groupFilter: number;
+  groupTag: number;
+  groupppl: number;
+  joinppl: number;
+
+  // store
+  storeName: string;
+  storeFilter: number;
+  storeTag: number;
+};
