+++ modules/LandingPage/data.tsx
@@ -74,52 +74,3 @@ export const defaultCommentsData: CommentDataType[] = [
     score: 0,
   },
 ];
-
-export const defaultGroupsData: GroupDataType[] = [
-  {
-    groupId: 0,
-    groupName: "",
-    groupStatus: "開團中",
-    place: "",
-    isPrivate: false,
-    store: null,
-    date: "",
-    startTime: "",
-    endTime: "",
-    totalMemberNum: 0,
-    currentpeople: 0,
-    isHomeGroup: false,
-    games: [
-      {
-        gameName: "",
-        gameType: 0,
-      },
-    ],
-
-    leader: {
-      userId: 0,
-      userName: "",
-      status: "member",
-      initNum: 1,
-      profileImg: "",
-    },
-    members: [],
-    tags: [],
-  },
-];
-
-export const defaultStoresData: StoreDataType[] = [
-  {
-    storeId: 0,
-    storeName: "",
-    address: "",
-    profileImg: "",
-    cover: "",
-    openHours: "",
-    score: 0,
-    description: "",
-    tags: [],
-    hqTag: true,
-    popTag: true,
-  },
-];
