+++ modules/SearchResults/FilterBlock.tsx
@@ -1,0 +1,31 @@
+import React from "react";
+import { FilterSetItemType } from "./data";
+
+export default function FilterBlock({ title, options }: FilterSetItemType) {
+  const textNum = title.length;
+
+  const spaces: Record<number, string> = {
+    2: "pl-11",
+    3: "pl-14.5",
+    4: "pl-18",
+    5: "pl-22.5",
+  };
+
+  const spaceStyle = spaces[textNum];
+
+  return (
+    <>
+      <label className="relative flex items-center flex-shrink-0 whitespace-nowrap text-sm">
+        <select
+          className={`min-h-9 rounded-sm bg-yellow-tint border-2 border-gray-500 text-gray-600 pr-3 py-1.5 ${spaceStyle}`}
+        >
+          {options.map((option) => {
+            const { value, text, checked } = option;
+            return <option key={value}>{text}</option>;
+          })}
+        </select>
+        <p className="absolute top-[7px] left-3 font-semibold">{title}</p>
+      </label>
+    </>
+  );
+}
